requirements.txt.txt
Flask
feedparser
beautifulsoup4
python-dateutil
cachetools
gunicorn
